src/povstorm_client/pyproject.toml
[tool.poetry]
name = "povstorm-client"
version = "0.1.0"
description = ""
authors = ["Bryce Lobdell <lobdellb@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.10.4"
pydantic-settings = "^2.7.1"
google-cloud-storage = "^2.19.0"
google-cloud-logging = "^3.11.3"
google-cloud-pubsub = "^2.27.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
